requirements.txt
kubernetes==23.6.0
cryptography==41.0.1
ldap3==2.9.1
PyYAML
prometheus-client==0.17.0
